chapter_003/src/test/java/ru/job4j/set/FullSearchTest.java
package ru.job4j.set;

import org.junit.Test;

import java.util.HashSet;
import java.util.List;
import java.util.Set;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

/**
 * Test.
 *
 * @author Yury Doronin (doronin.ltd@gmail.com)
 * @version 2.0
 * @since 05.01.2020
 */
public class FullSearchTest {

    /**
     * Test extractNumber.
     */
    @Test
    public void extractNumber() {
        List<Task> tasks = List.of(
                new Task("1", "First desc"),
                new Task("2", "Second desc"),
                new Task("1", "First desc"),
                new Task("1", "First desc")
        );
        Set<String> expected = new HashSet<>(List.of("1", "2"));
        assertThat(FullSearch.extractNumber1(tasks), is(expected));
        assertThat(FullSearch.extractNumber2(tasks), is(expected));
    }
}